refactor(frontend): move the theme provider into the design system
The application's provider mixed two concerns: deciding Infrahub's theme
policy, and wiring the machinery that carries it out. Only the first is
Infrahub's.

`@infrahub/ui` now exports a `ThemeProvider` that resolves a stored choice,
paints it, mirrors it for the next pre-paint, and fills the theme context so
`ThemeSwitchMenuItem` works anywhere below it. It takes `canChoose` and
`defaultTheme` and holds no opinion beyond them. The application's provider
keeps only the decision — read the flag, default a fresh user to dark — and is
now six lines of policy instead of thirty of plumbing.

It could not have moved wholesale: it calls `useConfig()`, so the design system
would have had to know Infrahub's config endpoint, its flag names and its query
client, inverting the dependency the package exists to avoid.

Also deletes `resolve-theme.ts` and the `ThemeChoice` model. They had no
consumer outside their own test — which is why knip stayed quiet — and worse,
`resolveTheme` answered "what theme?" with the opposite default to the code
that actually ships (light for an absent choice, where the provider gives
dark). Whoever wires up the account-level preference would have found it and
inherited the wrong behaviour. The vocabulary and its resolution rules are
recorded in `dev/specs/infp-46-dark-theme-completion/data-model.md`, and the
enum will arrive as a generated GraphQL type when that work lands.
