+++ frontend/packages/ui/src/index.ts
@@ -95,5 +95,6 @@ export {
 export { DismissGuardContext, useDismissGuard } from "./hooks/use-dissmiss-guard";
 export { applyTheme, type ResolvedTheme, useResolvedTheme } from "./theme/resolved-theme";
 export { ThemeContext, type ThemeControl, useThemeControl } from "./theme/theme-context";
+export { ThemeProvider, type ThemeProviderProps } from "./theme/theme-provider";
 export { mirrorResolvedTheme, readStoredChoice, storeChoice } from "./theme/theme-storage";
 export { ThemeSwitchMenuItem } from "./theme/theme-switch-menu-item";
